pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "musicorg"
version = "0.1.0"
description = "PySide6 desktop UI for beets music library management"
requires-python = ">=3.10"
dependencies = [
    "PySide6>=6.6.0",
    "beets>=2.0.0",
    "python3-discogs-client>=2.3.15",
    "mutagen>=1.47.0",
    "PyYAML>=6.0",
]

[project.scripts]
musicorg = "musicorg.__main__:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
